% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_import_from.R
\name{check_import_from}
\alias{check_import_from}
\title{Check \code{importFrom} in \code{NAMESPACE}}
\usage{
check_import_from(path = ".")
}
\arguments{
\item{path}{Path to package directory.}
}
\value{
Functions not used invisibly.
}
\description{
Checks if the functions declared in the NAMESPACE 
importFrom are actually used in the package functions.
}
\examples{
\dontrun{

# Execute in a package directory
check_import_from()

}
}
