% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-call.R
\name{count_calls}
\alias{count_calls}
\title{Count calls in a script or a directory}
\usage{
count_calls(path, ignore_base = TRUE, find_pkgs = TRUE)
}
\arguments{
\item{path}{Path to an R script or a directory, in that case all R scripts will be scanned.}

\item{ignore_base}{Ignore calls from base package.}

\item{find_pkgs}{Try to find from which package calls are from, need to scan for dependencies with \code{renv}.}
}
\value{
a \code{data.frame} with 2 columns: \code{call} and \code{n}
}
\description{
Count calls in a script or a directory
}
\examples{
\dontrun{

count_calls("PATH/TO/SCRIPTS")

}
}
