% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixer.R
\name{prefixer}
\alias{prefixer}
\title{Shiny Gadget to interactively prefix function in a script}
\usage{
prefixer()
}
\description{
Open a script, load package(s) used in that script, then launch this addin via RStudio menu.
It will propose to prefix function with their package (if a function is in several packages, a choice can be made),
validate the suggested prefix or skip to the next function.
}
\examples{
\dontrun{

prefixer()

}
}
