% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_nonascii_file.R
\name{show_nonascii_file}
\alias{show_nonascii_file}
\title{Search for non-ASCII characters in files}
\usage{
show_nonascii_file(
  path = ".",
  size_limit = 5e+05,
  ignore_dir = c("docs"),
  ignore_ext = c("png", "jpg", "rds", "rda")
)
}
\arguments{
\item{path}{Path where to start searching. If a directory,
search in all files recursively.}

\item{size_limit}{Limit size of files scanned (in bytes).}

\item{ignore_dir}{Ignore some directories (relative to \code{path}).}

\item{ignore_ext}{Ignore with those extensions.}
}
\value{
a \code{data.frame} with files and lines where
 non-ascii have been detected, if none return \code{NULL}.
}
\description{
Search for non-ASCII characters in files
}
\examples{

\dontrun{

show_nonascii_file(system.file(package = "prefixer"))

}

}
