% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Rd.R
\name{check-rd}
\alias{check-rd}
\alias{check_Rd_examples}
\alias{check_Rd_value}
\title{Check Rd file}
\usage{
check_Rd_examples(path = "man/")

check_Rd_value(path = "man/")
}
\arguments{
\item{path}{Path to Rd files, typically \code{man/}}
}
\value{
Character vector of Rd files path.
}
\description{
Check Rd file
}
\examples{
\dontrun{

# Check if Rd files have examples
check_Rd_examples()

# Check if Rd files have value (function return)
check_Rd_value()

}
}
