% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-from.R
\name{import_from}
\alias{import_from}
\title{Generate importFrom tag from a function}
\usage{
import_from(fun)
}
\arguments{
\item{fun}{A function.}
}
\value{
Character vector
}
\description{
Works only if functions used are prefixed in the body.
}
\examples{

my_fun <- function(path) {
  read.table(file = path, header = FALSE, sep = "\\t")
}
import_from(my_fun)

}
