% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_nonascii_file.R
\name{show_nonascii_file}
\alias{show_nonascii_file}
\title{Search for non-ASCII characters in files}
\usage{
show_nonascii_file(path = ".")
}
\arguments{
\item{path}{Path where to start searching. If a directory,
search in all files recursively.}
}
\value{
a \code{data.frame} with files and lines where
 non-ascii have been detected, if none return \code{NULL}.
}
\description{
Search for non-ASCII characters in files
}
\examples{

\dontrun{

show_nonascii_file(system.file(package = "prefixer"))

}

}
